crop_disease_predictor-main/requirements.txt
tensorflow==2.15.0
numpy==1.26.4
flask==3.0.0
werkzeug==3.0.1
pillow==10.0.0
scikit-learn==1.3.2
gunicorn==21.2.0
